% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate_page}
\alias{translate_page}
\title{Function translating one placeholder to actual content}
\usage{
translate_page(page_path, language, glossary = NULL)
}
\arguments{
\item{page_path}{Path to the placeholder. First create on main,
for instance to translate index.Rmd to Portuguese,
first create a copy of index.Rmd as index.pt.Rmd on main.}

\item{language}{Language to which to translate.
See \href{https://www.deepl.com/docs-api/translate-text}{docs of the \code{source_lang} and \code{target_lang} API parameters}.}

\item{glossary}{Glossary name.}
}
\value{
Nothing
}
\description{
Also helps opening the PR
}
\details{
First follow babeldown docs to set the two environment variables.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{Sys.setenv("DEEPL_API_URL" = "https://api.deepl.com")
Sys.setenv(DEEPL_API_KEY = keyring::key_get("deepl"))
}\if{html}{\out{</div>}}
}
\examples{
translate_page("index.pt.Rmd", "PT-BR")
}
